pyproject.toml
[tool.poetry]
name = "django-common-server"
version = "0.0.1"
description = "django-common-server"
authors = ["paullee714 <paul_lee@kakao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
